4542 | 118 files | M Protobufs/econ_gcmessages.proto,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/bin/linuxsteamrt64/libserver_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/pak01_dir/cfg/video_defaults_0.txt,  M game/dota/pak01_dir/cfg/video_defaults_1.txt,  M game/dota/pak01_dir/cfg/video_defaults_2.txt,  M game/dota/pak01_dir/cfg/video_defaults_3.txt,  M game/dota/pak01_dir/panorama/styles/dashboard.css,  M game/dota/pak01_dir/panorama/styles/ui_econ_item_diretide.css,  M game/dota/pak01_dir/resource/localization/abilities_brazilian.txt,  M game/dota/pak01_dir/resource/localization/abilities_bulgarian.txt,  M game/dota/pak01_dir/resource/localization/abilities_czech.txt,  M game/dota/pak01_dir/resource/localization/abilities_danish.txt,  M game/dota/pak01_dir/resource/localization/abilities_dutch.txt,  M game/dota/pak01_dir/resource/localization/abilities_finnish.txt,  M game/dota/pak01_dir/resource/localization/abilities_french.txt,  M game/dota/pak01_dir/resourc
https://steamdb.info/patchnotes/5795588/

+++ game/dota_addons/diretide2020/scripts/vscripts/addon_game_mode.lua
@@ -1284,6 +1284,9 @@ function CDiretide:EndRound( nScoringTeam )
 			Hero:RemoveModifierByName( "modifier_wisp_spirits" )
 			Hero:RemoveModifierByName( "modifier_beastmaster_wild_axes" )
 			Hero:RemoveModifierByName( "modifier_life_stealer_infest" )
+			Hero:RemoveModifierByName( "modifier_pudge_rot" )
+			Hero:RemoveModifierByName( "modifier_kunkka_x_marks_the_spot_marker" )
+			Hero:RemoveModifierByName( "modifier_kunkka_x_marks_the_spot" )
 
 			Hero:AddNewModifier( Hero, nil, "modifier_hero_post_round", { duration = 2.0 } )
 
@@ -1305,7 +1308,15 @@ function CDiretide:EndRound( nScoringTeam )
 	local vecThinkers = Entities:FindAllByClassname( "npc_dota_thinker" )
 	if vecThinkers ~= nil then
 		for _, hThinker in ipairs( vecThinkers ) do
-			hThinker:RemoveModifierByName( "modifier_dark_seer_wall_of_replica" ) -- will kill thinker since the modifier's OnDestroy does so.
+			if hThinker ~= nil and hThinker:IsNull() == false then
+				local hBuff = hThinker:FindModifierByName( "modifier_kunkka_x_marks_the_spot_thinker" )
+				if hBuff ~= nil and hBuff:IsNull() == false then
+					hBuff:GetCaster():UnHideAbilityToSlot( "kunkka_x_marks_the_spot", "kunkka_return" )
+					UTIL_Remove( hThinker )
+				else
+					hThinker:RemoveModifierByName( "modifier_dark_seer_wall_of_replica" ) -- will kill thinker since the modifier's OnDestroy does so.
+				end
+			end
 		end
 	end
 
@@ -1564,6 +1575,9 @@ function CDiretide:ResetRoshan()
 
 	self.nQueueRoshanForTeam = nil
 
+	-- JUST IN CASE let's kill the soundloop on all players.
+	self:KillRoshanChaseSound()
+
 	self.hRoshan:SetAbsAngles( 0, 270, 0 )
 	self.hRoshan.nCandyAttackTeam = nil
 	self.hRoshan.nTrickOrTreatTeam = 0
@@ -1621,8 +1635,23 @@ end
 
 --------------------------------------------------------------------------------
 
+function CDiretide:KillRoshanChaseSound()
+	-- JUST IN CASE let's kill the soundloop on all players.
+	for nPlayerID = 0, DOTA_MAX_PLAYERS-1 do
+		local hPlayer = PlayerResource:GetPlayer( nPlayerID )
+		if hPlayer ~= nil then
+			StopSoundEvent( "RoshanTarget.Loop", hPlayer )
+		end
+	end
+end
+
+--------------------------------------------------------------------------------
+
 function CDiretide:TrickOrTreatToTeam( nTeam, bIncrementCounter )
 	self.hRoshan.nTrickOrTreatTeam = 0 -- we'll set this later, if we succeed
+
+	-- JUST IN CASE let's kill the soundloop on all players.
+	self:KillRoshanChaseSound()
 
 	-- catch running this during round ending
 	if self:IsRoundInProgress() == false then
